Make it a monorepo and convert demos to TypeScript
