+++ demos/console-demo/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "console-demo",
+  "type": "module",
+  "version": "1.0.0",
+  "scripts": {
+    "start": "npm run build && node build/demo.js",
+    "build": "tsc"
+  },
+  "dependencies": {
+    "jsondiffpatch": "^0.5.0"
+  },
+  "devDependencies": {
+    "typescript": "~5.2.2"
+  }
+}
