+++ demos/html-demo/index.html
@@ -26,10 +26,6 @@
     />
     <script
       type="text/javascript"
-      src="https://cdn.jsdelivr.net/npm/jsondiffpatch/dist/jsondiffpatch.umd.min.js"
-    ></script>
-    <script
-      type="text/javascript"
       src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/3.21.0/codemirror.js"
     ></script>
     <script
@@ -187,7 +183,7 @@
         <a href="https://twitter.com/beneidel">Benjamin Eidelman</a>
       </p>
     </footer>
-    <script type="text/javascript" src="demo.js"></script>
+    <script src="build/demo.js" type="module"></script>
 
     <script>
       (function (i, s, o, g, r, a, m) {
