+++ demos/html-demo/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "html-demo",
+  "type": "module",
+  "version": "1.0.0",
+  "scripts": {
+    "build": "tsc && esbuild demo.ts --bundle --outdir=build"
+  },
+  "dependencies": {
+    "jsondiffpatch": "^0.5.0"
+  },
+  "devDependencies": {
+    "esbuild": "^0.19.8",
+    "typescript": "~5.2.2"
+  }
+}
