+++ demos/numeric-plugin-demo/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "numeric-plugin-demo",
+  "type": "module",
+  "version": "1.0.0",
+  "scripts": {
+    "start": "npm run build && node build/numeric-plugin.js",
+    "build": "tsc"
+  },
+  "dependencies": {
+    "jsondiffpatch": "^0.5.0"
+  },
+  "devDependencies": {
+    "typescript": "~5.2.2"
+  }
+}
