+++ package.json
@@ -1,70 +1,16 @@
 {
   "name": "jsondiffpatch",
-  "version": "0.5.0",
-  "author": "Benjamin Eidelman <beneidel@gmail.com>",
-  "description": "Diff & Patch for Javascript objects",
-  "contributors": [
-    "Benjamin Eidelman <beneidel@gmail.com>"
-  ],
-  "browser": "dist/jsondiffpatch.umd.js",
-  "main": "dist/jsondiffpatch.cjs.js",
-  "module": "dist/jsondiffpatch.esm.js",
-  "types": "dist/types/main.d.ts",
-  "files": [
-    "dist",
-    "bin"
-  ],
-  "bin": {
-    "jsondiffpatch": "./bin/jsondiffpatch"
-  },
   "scripts": {
-    "build": "rollup -c",
-    "type-check": "tsc",
-    "lint": "eslint . --ext .ts",
-    "test": "jest --coverage",
     "format": "prettier . --write",
-    "format-check": "prettier . --check",
-    "prepack": "npm run build",
-    "prepublishOnly": "npm run test && npm run lint"
+    "format-check": "prettier . --check"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/benjamine/jsondiffpatch.git"
-  },
-  "keywords": [
-    "json",
-    "diff",
-    "patch"
-  ],
   "dependencies": {
-    "chalk": "^3.0.0",
-    "diff-match-patch": "^1.0.5"
+    "prettier": "^3.0.3"
   },
-  "devDependencies": {
-    "@babel/core": "^7.22.15",
-    "@babel/preset-env": "^7.22.15",
-    "@babel/preset-typescript": "^7.22.15",
-    "@rollup/plugin-babel": "^6.0.3",
-    "@rollup/plugin-commonjs": "^25.0.4",
-    "@rollup/plugin-node-resolve": "^15.2.0",
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-typescript": "^11.1.3",
-    "@types/jest": "^29.5.4",
-    "@typescript-eslint/eslint-plugin": "^6.6.0",
-    "@typescript-eslint/parser": "^6.6.0",
-    "eslint": "^8.47.0",
-    "eslint-config-prettier": "^9.0.0",
-    "jest": "^29.6.2",
-    "prettier": "^3.0.2",
-    "rollup": "^3.28.0",
-    "rollup-plugin-visualizer": "^5.9.2",
-    "ts-jest": "^29.1.1",
-    "tslib": "^2.6.2",
-    "typescript": "~5.2.2"
-  },
-  "license": "MIT",
-  "engines": {
-    "node": ">=8.17.0"
-  },
-  "homepage": "https://github.com/benjamine/jsondiffpatch"
+  "workspaces": [
+    "demos/console-demo",
+    "demos/html-demo",
+    "demos/numeric-plugin-demo",
+    "packages/jsondiffpatch"
+  ]
 }
