+++ packages/jsondiffpatch/.babelrc
