+++ packages/jsondiffpatch/.editorconfig
