+++ packages/jsondiffpatch/.eslintignore
