+++ packages/jsondiffpatch/.eslintrc.cjs
