+++ packages/jsondiffpatch/src/filters/lcs.ts
