+++ packages/jsondiffpatch/src/filters/texts.ts
