+++ packages/jsondiffpatch/src/filters/trivial.ts
