+++ packages/jsondiffpatch/src/formatters/annotated.ts
