+++ packages/jsondiffpatch/src/formatters/console.ts
