+++ packages/jsondiffpatch/src/formatters/styles/annotated.css
