+++ packages/jsondiffpatch/src/pipe.ts
